resume_analyser_AI-main/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0